requirements.txt
# Core requirements
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.31.1
pydantic==2.5.3
tensorflow==2.15.0
numpy==1.26.3
pillow==10.2.0
google-generativeai==0.5.0
opencv-python==4.8.1.78
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.1
